Implement DISPLAY_DEBUG flag
This flag uses InkVT (from NiLuJe's fork: https://github.com/NiLuJe/inkvt , thanks to @llandsmeer too) to display the entire output of the rcS on the eInk screen when it usually is on the serial console. It is available for both standard and rooted kernels, and can be enabled by creating a file named "DISPLAY_DEBUG" and with "true" as its contents. Likewise, it can be disabled by writing "false" into it.
It can be used to provide debug information or diagnose problems more easily without needing to connect to a serial console.