`startx`: Some fixes
Since the now-default platform plugins that InkBox GUI uses are linked to `libinput`, this seems to create an unexplainable panic from the VNC touch input handler which just throws `Segmentation fault` for no apparent reason. Fixing this for now by mounting in-place the old platform plugins which weren't linked to `libinput` et. al only in this situation. The rest of the system still uses the `libinput` ones where possible.